# contract/openapi.yaml
...
openapi: 3.1.0
info:
  title: pantry recipe-service
  version: "1.0.0"
  description: >-
    The canonical source of truth for recipe definitions, ingredient
    normalization, grocery-list aggregation, nutrition, pricing and
    recommendations. It holds no user state: Convex passes the full user context
    on every request.
servers:
  - url: http://recipe-service:8080
    description: The compose-network address Convex actions call.
security:
  - serviceSecret: []
    userId: []
...
paths:
...
  /pricing/estimate:
    post:
      tags: [pricing]
      operationId: pricingEstimate
      summary: Price a set of already-aggregated grocery lines.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PricingEstimateRequest" }
      responses:
        "200":
          description: The estimate, its per-line outcomes, and its basis.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CostEstimate" }
        "500": { $ref: "#/components/responses/ServerError" }
...
  /pricing/stores:
    post:
      tags: [pricing]
      operationId: pricingStores
      summary: Find stores near a zip code for the user to price against.
      description: >-
        Returns an empty list and no provider when real store prices are not
        configured. A 502 means the retailer's directory could not be reached,
        which is a failure of this search alone — pricing a grocery list never
        fails, it falls back to the national averages.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StoreSearchRequest" }
      responses:
        "200":
          description: The nearby stores, or an empty list when the feature is off.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StoreSearchResult" }
        "502": { $ref: "#/components/responses/ServerError" }
...
components:
  securitySchemes:
    serviceSecret:
      type: apiKey
      in: header
      name: X-Service-Secret
      description: The shared secret every route but /healthz is gated behind.
    userId:
      type: apiKey
      in: header
      name: X-User-Id
      description: >-
        The authenticated user, supplied by Convex. recipe-service holds no user
        state, so this is the only thing that scopes a request.
...
  responses:
...
    ServerError:
      description: The service could not complete the request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: |-
        Every non-2xx response body. One field, always present: the message
        `writeError` recorded, never a stack or an internal cause.
      x-go-unbound: written inline by writeErr as a map literal.
      required: [error]
      properties:
        error: { type: string }
...
    PriceStaleness:
      type: string
      description: How stale the underlying price table is, graded when the estimate is made.
      x-go-unbound: pricing.Staleness is a named string type, checked as a string.
      enum: [fresh, aging, stale]
    PriceBasis:
      type: object
      description: >-
        Provenance for an estimate. A total shown without this is a number
        pretending to be a price.
      x-go-types: [pricing.Basis]
      required: [source, sourceUrl, area, observationMonth, staleness]
      properties:
        source: { type: string }
        sourceUrl: { type: string }
        area:
          description: Geographic basis of the averages, e.g. "U.S. city average".
          type: string
        observationMonth:
          description: '"YYYY-MM" the prices were observed.'
          type: string
        staleness: { $ref: "#/components/schemas/PriceStaleness" }
        store: { $ref: "#/components/schemas/StorePriceBasis" }
    PricedLine:
      type: object
      description: >-
        Per-line outcome. Unpriced lines carry a reason and contribute nothing to
        the total.
      x-go-types: [pricing.LineEstimate]
      required: [canonicalItem, item, priced]
      properties:
        canonicalItem: { type: string }
        item: { type: string }
        priced: { type: boolean }
        cents: { type: integer }
        bucket:
          description: Which coarse price bucket the ingredient resolved to.
          type: string
        bucketLabel: { type: string }
        reason:
          description: Why the line could not be priced, when `priced` is false.
          type: string
        source: { $ref: "#/components/schemas/PriceSource" }
        product:
          description: >-
            The shelf item a store price came from, so a number can be shown as
            the price of something specific. Absent for an average.
          type: string
        onSale:
          description: >-
            True when the store price is a promotion rather than the regular
            shelf price, and so will not last.
          type: boolean
    CostEstimate:
      type: object
      description: |-
        Grocery pricing (BL-0023 increment 1).

        Deliberately a separate contract from GroceryLine rather than fields on
        it: price data has its own source, refresh cadence and legal constraints,
        and the grocery list must keep working when pricing is unavailable.
      x-go-types: [pricing.Estimate]
      required: [currency, totalCents, pricedCount, unpricedCount, lines, basis]
      properties:
        currency: { type: string }
        totalCents: { type: integer }
        pricedCount: { type: integer }
        unpricedCount: { type: integer }
        lines:
          type: array
          items: { $ref: "#/components/schemas/PricedLine" }
        basis: { $ref: "#/components/schemas/PriceBasis" }
    PricingLine:
      type: object
      description: >-
        One aggregated grocery line to price: the normalized identity plus its
        quantity.
      x-go-types: [pricing.Line]
      x-go-direction: request
      required: [canonicalItem, item, unit, quantity]
      properties:
        canonicalItem: { type: string }
        item: { type: string }
        unit: { type: string }
        quantity: { type: number }
    PricingEstimateRequest:
      type: object
      x-go-unbound: decoded into an anonymous struct in the pricingEstimate handler.
      x-go-direction: request
      required: [lines]
      properties:
        lines:
          type: array
          items: { $ref: "#/components/schemas/PricingLine" }
        storeLocationId:
          description: >-
            The store the user opted into (BL-0046). Absent — which is every
            user who has not chosen one — prices the list from the national
            averages alone.
          type: string
    PriceSource:
      type: string
      description: >-
        Which table priced a line. Absent on an unpriced line, where there is no
        number to attribute.
      x-go-unbound: a plain string field on pricing.LineEstimate, checked as a string.
      enum: [average, store]
    StorePriceBasis:
      type: object
      description: |-
        Provenance for the store-priced half of an estimate (BL-0046).

        It sits alongside the averages basis rather than replacing it, because a
        mixed estimate genuinely has two sources, and is present only when the
        chosen store actually priced something.
      x-go-types: [pricing.StoreBasis]
      required: [provider, locationId, fetchedAt, pricedCount]
      properties:
        provider: { type: string }
        locationId: { type: string }
        storeName: { type: string }
        fetchedAt:
          description: RFC 3339. Shelf prices move daily, so the age of the quote is part of the number.
          type: string
        pricedCount:
          description: >-
            How many lines the store priced. The rest came from the averages,
            and saying so is what keeps a mixed total honest.
          type: integer
    StoreLocation:
      type: object
      description: One store a user can choose to price against.
      x-go-types: [pricing.StoreLocation]
      required: [locationId, name]
      properties:
        locationId: { type: string }
        name: { type: string }
        chain: { type: string }
        address: { type: string }
        city: { type: string }
        state: { type: string }
        zipCode: { type: string }
    StoreSearchRequest:
      type: object
      x-go-unbound: decoded into an anonymous struct in the pricingStores handler.
      x-go-direction: request
      required: [zipCode]
      properties:
        zipCode: { type: string }
        radiusMiles:
          description: Defaults to 10 and is capped at 100 by the provider.
          type: integer
    StoreSearchResult:
      type: object
      description: >-
        Nearby stores. An empty provider means real store prices are not
        configured, which a client renders by offering no chooser at all.
      x-go-types: [pricing.StoreSearch]
      required: [provider, stores]
      properties:
        provider: { type: string }
        stores:
          type: array
          items: { $ref: "#/components/schemas/StoreLocation" }
...